LayoutTests/pointerevents/ios/pointer-event-twist-on-stylus-down.html

<!DOCTYPE html>
<html>
<head>
    <script src="../../resources/js-test.js"></script>
    <script src="../../resources/ui-helper.js"></script>
</head>
<script>
    jsTestIsAsync = true;

    if (window.testRunner) {
        testRunner.waitUntilDone();
        testRunner.dumpAsText();

        document.addEventListener("pointerdown", function (e) {
            twist = e.twist;
            shouldBe("twist", "0");
            finishJSTest();
        });

        UIHelper.stylusTapAt(10, 10);
    }
</script>
</html>
