LayoutTests/pointerevents/ios/pointer-event-twist-on-stylus-down-expected.txt

PASS twist is 0
PASS successfullyParsed is true

TEST COMPLETE
